This is nearing completion
The game is almost done at this point, with the end routines and final battle needing to be coded. After this it will just be some tweaks. As this was never intended to be more than  project to see if I could do it, I am not sure I will polish it more after completion.