Final Upload (I think)
This is as far as I am willing to take the game to this point. If someone else want to fix the remaining bugs they find, that is up to them. The goal was to prove and complete a text adventure game, and I have done that. I have also learned ways to do things better and definitely should stream line code, which I will do in the future.